requirements.txt
mrs-mapreduce